Fixed bug with crash when application with opened dialog moves to background. This has happened because the listener serialization.
