+++ .idea/modules.xml
@@ -3,7 +3,10 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/AndroidFileChooser.iml" filepath="$PROJECT_DIR$/AndroidFileChooser.iml" />
+      <module fileurl="file://$PROJECT_DIR$/AndroidFileChooser.iml" filepath="$PROJECT_DIR$/AndroidFileChooser.iml" />
       <module fileurl="file://$PROJECT_DIR$/android-file-chooser/android-file-chooser.iml" filepath="$PROJECT_DIR$/android-file-chooser/android-file-chooser.iml" />
+      <module fileurl="file://$PROJECT_DIR$/android-file-chooser/android-file-chooser.iml" filepath="$PROJECT_DIR$/android-file-chooser/android-file-chooser.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
