+++ android-file-chooser/src/main/java/ir/sohreco/androidfilechooser/ExternalStorageNotAvailableException.java
@@ -1,7 +1,7 @@
 package ir.sohreco.androidfilechooser;
 
-public class ExternalStorageNotAvailableException extends Exception {
-    public ExternalStorageNotAvailableException() {
+class ExternalStorageNotAvailableException extends RuntimeException {
+    ExternalStorageNotAvailableException() {
         super("There is no external storage available on this device.");
     }
 }
