+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="ir.sohreco.androidfilechooser.MainActivity">
+    tools:context="ir.sohreco.androidfilechooser.sample.MainActivity">
 
     <TextView
         android:layout_width="wrap_content"
