+++ app/src/test/java/ir/sohreco/androidfilechooser/sample/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package ir.sohreco.androidfilechooser;
+package ir.sohreco.androidfilechooser.sample;
 
 import org.junit.Test;
